% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeReg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{ridgereg Class}
\value{
ridgereg object
}
\description{
ridgereg Class
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric.}

\item{\code{beta}}{numeric.}

\item{\code{dataNames}}{character.}

\item{\code{yHat}}{numeric.}
}}


\examples{
data(iris)
ridgereg <-  ridgereg(Petal.Length ~ Species, data = iris)
ridgereg$print()
ridgereg$predict()
ridgereg$coef()
}
