% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Airport_Delay.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Airport Delay Visualise on a ggplot}
\usage{
visualize_airport_delays()
}
\value{
ggplot about airport delays
}
\description{
Airport Delay Visualise on a ggplot
}
\examples{
visualize_airport_delays()
}
